requirements.txt
aiolimiter==1.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
cryptography==41.0.7
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
psycopg==3.1.17
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.25
tornado==6.3.3
typing_extensions==4.9.0
tzlocal==5.2
wheel==0.42.0
